Reorganized barChart functions (#58)
* Added barChart component. Works for one and two weeks. Also created basic layout for statistics page

* added component for a dropdown menu that contains alternatives for intervals

* Added barChart + fixed style

* Added basic layout to statistics page

* Changed language of name of months from english to swedish

* Fixed bug with requests not working cause of bad query

* Hopefully fixed bug with promises (needed catch)

* Fixed changes based on feedback in GitHub. Some minor errors in react

* Minor changes. Again

* Removed empty file

* Moved some functions from statistics.js to barChart.js

* Removed comments and console.log():s

* Removed unnecessary code and fixed style issues

* Added missing apostrophe in statistics.js

* Style fixed
