+++ frontend-friends/src/routes/statistics/Statistics.scss
@@ -8,9 +8,8 @@
     }
 
     .chart-container{
-        //width: 100%;
         height: 100%;
-        display: grid;
+        display: grid; 
         grid-template-columns: repeat(3, 390px);
         background-color: white;
         margin-left: 40px;
